requirements.txt
aiocron
openai
beautifulsoup4
selenium
pandas
aiohttp
feedparser
ruamel.yaml
qq-botpy